remove props files
